continue bug fixing
