+++ src/pages/music-collections/categories-of-hits.js
@@ -7,18 +7,14 @@ import * as S from "../../components/playlist/playlist.styled";
 export const CategoriesOfHits = ({ setTitle, searchText }) => {
   const params = useParams();
   const selectionId = parseInt(params.id);
-  // console.log("-->", params);
-  // console.log(typeof selectionId);
-  // console.log(selectionId);
+
   const category = categories.find((category) => category.id === selectionId);
-  // console.log(category);
 
   setTitle(category.selection_title);
 
   const { data, error, isLoading } = useGetSelectionByIdQuery({
     id: selectionId,
   });
-  console.log(data);
 
   const isEmptyList = !isLoading && (!data || data.length === 0);
   if (isEmptyList) {
