+++ src/pages/my-playlist/my-playlist.jsx
@@ -1,4 +1,4 @@
-import { useGetMyTracksQuery } from "../../services/api-services";
+import { useGetMyTracksQuery } from "../../services/api-services-reauth";
 import { TrackListComponent } from "../main-page/main-page";
 import * as S from "../../components/playlist/playlist.styled";
 import { CurrentUserContext } from "../../routes";
@@ -14,6 +14,7 @@ export const MyPlaylist = ({ isVisiable, searchText }) => {
   }
 
   if (error) {
+    
     if (error.status === 401) {
       handleLogout();
     }
