+++ src/store/store.js
@@ -1,6 +1,7 @@
 import { configureStore } from "@reduxjs/toolkit";
 import trackSlice from "./slices/slices";
-import { allTracksApi, myTracksApi, token } from "../services/api-services";
+import { allTracksApi, token } from "../services/api-services";
+import { myTracksApi } from "../services/api-services-reauth";
 
 export const store = configureStore({
   reducer: {
